condition render account route using customerAccessToken
