+++ examples/template-hydrogen-default/src/routes/api/login.server.js
@@ -3,18 +3,6 @@ import cookie from 'cookie';
 
 import {CUSTOMER_ACCESS_TOKEN_COOKIE_NAME} from '../../constants/cookies';
 
-import Layout from '../../components/Layout.server';
-import {LoginForm} from '../../components/LoginForm.client';
-
-export default function Login() {
-  return (
-    <Layout>
-      <h1 className="text-2xl font-bold">Login</h1>
-      <LoginForm />
-    </Layout>
-  );
-}
-
 export async function api(request, {queryShop}) {
   const jsonBody = await request.json();
